Upgrade to Scala 2.13.3 & Scala.js 1.1.1

+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/Types.scala
@@ -1,11 +1,9 @@
 package io.github.shogowada.scala.jsonrpc
 
-import io.github.shogowada.scala.jsonrpc.serializers.JSONRPCPickler.{ReadWriter => RW, macroRW, readwriter}
-
 import scala.concurrent.Future
 
 object Types {
-  type Id = Either[String, Integer]
+  type Id = Either[String, BigDecimal]
 
   type JSONSender = (String) => Future[Option[String]]
 }
